chore: extract lib/hooks.ts file
