+++ packages/react-router/index.tsx
@@ -22,25 +22,31 @@ import {
   NavigationContext,
   RouteContext,
 } from "./lib/context";
+import type { Params, PathMatch, RouteMatch, RouteObject } from "./lib/router";
+import type { NavigateFunction, NavigateOptions } from "./lib/hooks";
+import {
+  _renderMatches,
+  useHref,
+  useInRouterContext,
+  useLocation,
+  useMatch,
+  useNavigationType,
+  useNavigate,
+  useOutlet,
+  useOutletContext,
+  useParams,
+  useResolvedPath,
+  useRoutes,
+} from "./lib/hooks";
 import {
   generatePath,
-  getToPathname,
   invariant,
-  joinPaths,
   matchPath,
   matchRoutes,
   normalizePathname,
-  ParamParseKey,
-  Params,
-  PathMatch,
-  PathPattern,
   resolvePath,
-  resolveTo,
-  RouteMatch,
-  RouteObject,
   stripBasename,
   warning,
-  warningOnce,
 } from "./lib/router";
 
 // re-export from history
@@ -53,6 +59,22 @@ export type { Navigator };
 // re-export from lib/router.ts
 export type { Params, PathMatch, RouteMatch, RouteObject };
 export { generatePath, matchPath, matchRoutes, resolvePath };
+
+// re-export from lib/hooks.ts
+export type { NavigateFunction, NavigateOptions };
+export {
+  useHref,
+  useInRouterContext,
+  useLocation,
+  useMatch,
+  useNavigationType,
+  useNavigate,
+  useOutlet,
+  useOutletContext,
+  useParams,
+  useResolvedPath,
+  useRoutes,
+};
 
 ///////////////////////////////////////////////////////////////////////////////
 // COMPONENTS
@@ -298,362 +320,6 @@ export function Routes({
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-// HOOKS
-///////////////////////////////////////////////////////////////////////////////
-
-/**
- * Returns the full href for the given "to" value. This is useful for building
- * custom links that are also accessible and preserve right-click behavior.
- *
- * @see https://reactrouter.com/docs/en/v6/api#usehref
- */
-export function useHref(to: To): string {
-  invariant(
-    useInRouterContext(),
-    // TODO: This error is probably because they somehow have 2 versions of the
-    // router loaded. We can help them understand how to avoid that.
-    `useHref() may be used only in the context of a <Router> component.`
-  );
-
-  let { basename, navigator } = React.useContext(NavigationContext);
-  let { hash, pathname, search } = useResolvedPath(to);
-
-  let joinedPathname = pathname;
-  if (basename !== "/") {
-    let toPathname = getToPathname(to);
-    let endsWithSlash = toPathname != null && toPathname.endsWith("/");
-    joinedPathname =
-      pathname === "/"
-        ? basename + (endsWithSlash ? "/" : "")
-        : joinPaths([basename, pathname]);
-  }
-
-  return navigator.createHref({ pathname: joinedPathname, search, hash });
-}
-
-/**
- * Returns true if this component is a descendant of a <Router>.
- *
- * @see https://reactrouter.com/docs/en/v6/api#useinroutercontext
- */
-export function useInRouterContext(): boolean {
-  return React.useContext(LocationContext) != null;
-}
-
-/**
- * Returns the current location object, which represents the current URL in web
- * browsers.
- *
- * Note: If you're using this it may mean you're doing some of your own
- * "routing" in your app, and we'd like to know what your use case is. We may
- * be able to provide something higher-level to better suit your needs.
- *
- * @see https://reactrouter.com/docs/en/v6/api#uselocation
- */
-export function useLocation(): Location {
-  invariant(
-    useInRouterContext(),
-    // TODO: This error is probably because they somehow have 2 versions of the
-    // router loaded. We can help them understand how to avoid that.
-    `useLocation() may be used only in the context of a <Router> component.`
-  );
-
-  return React.useContext(LocationContext).location;
-}
-
-/**
- * Returns the current navigation action which describes how the router came to
- * the current location, either by a pop, push, or replace on the history stack.
- *
- * @see https://reactrouter.com/docs/en/v6/api#usenavigationtype
- */
-export function useNavigationType(): NavigationType {
-  return React.useContext(LocationContext).navigationType;
-}
-
-/**
- * Returns true if the URL for the given "to" value matches the current URL.
- * This is useful for components that need to know "active" state, e.g.
- * <NavLink>.
- *
- * @see https://reactrouter.com/docs/en/v6/api#usematch
- */
-export function useMatch<
-  ParamKey extends ParamParseKey<Path>,
-  Path extends string
->(pattern: PathPattern<Path> | Path): PathMatch<ParamKey> | null {
-  invariant(
-    useInRouterContext(),
-    // TODO: This error is probably because they somehow have 2 versions of the
-    // router loaded. We can help them understand how to avoid that.
-    `useMatch() may be used only in the context of a <Router> component.`
-  );
-
-  let { pathname } = useLocation();
-  return React.useMemo(
-    () => matchPath<ParamKey, Path>(pattern, pathname),
-    [pathname, pattern]
-  );
-}
-
-/**
- * The interface for the navigate() function returned from useNavigate().
- */
-export interface NavigateFunction {
-  (to: To, options?: NavigateOptions): void;
-  (delta: number): void;
-}
-
-export interface NavigateOptions {
-  replace?: boolean;
-  state?: any;
-}
-
-/**
- * Returns an imperative method for changing the location. Used by <Link>s, but
- * may also be used by other elements to change the location.
- *
- * @see https://reactrouter.com/docs/en/v6/api#usenavigate
- */
-export function useNavigate(): NavigateFunction {
-  invariant(
-    useInRouterContext(),
-    // TODO: This error is probably because they somehow have 2 versions of the
-    // router loaded. We can help them understand how to avoid that.
-    `useNavigate() may be used only in the context of a <Router> component.`
-  );
-
-  let { basename, navigator } = React.useContext(NavigationContext);
-  let { matches } = React.useContext(RouteContext);
-  let { pathname: locationPathname } = useLocation();
-
-  let routePathnamesJson = JSON.stringify(
-    matches.map((match) => match.pathnameBase)
-  );
-
-  let activeRef = React.useRef(false);
-  React.useEffect(() => {
-    activeRef.current = true;
-  });
-
-  let navigate: NavigateFunction = React.useCallback(
-    (to: To | number, options: NavigateOptions = {}) => {
-      warning(
-        activeRef.current,
-        `You should call navigate() in a React.useEffect(), not when ` +
-          `your component is first rendered.`
-      );
-
-      if (!activeRef.current) return;
-
-      if (typeof to === "number") {
-        navigator.go(to);
-        return;
-      }
-
-      let path = resolveTo(
-        to,
-        JSON.parse(routePathnamesJson),
-        locationPathname
-      );
-
-      if (basename !== "/") {
-        path.pathname = joinPaths([basename, path.pathname]);
-      }
-
-      (!!options.replace ? navigator.replace : navigator.push)(
-        path,
-        options.state
-      );
-    },
-    [basename, navigator, routePathnamesJson, locationPathname]
-  );
-
-  return navigate;
-}
-
-const OutletContext = React.createContext<unknown>(null);
-
-/**
- * Returns the context (if provided) for the child route at this level of the route
- * hierarchy.
- * @see https://reactrouter.com/docs/en/v6/api#useoutletcontext
- */
-export function useOutletContext<Context = unknown>(): Context {
-  return React.useContext(OutletContext) as Context;
-}
-
-/**
- * Returns the element for the child route at this level of the route
- * hierarchy. Used internally by <Outlet> to render child routes.
- *
- * @see https://reactrouter.com/docs/en/v6/api#useoutlet
- */
-export function useOutlet(context?: unknown): React.ReactElement | null {
-  let outlet = React.useContext(RouteContext).outlet;
-  if (outlet) {
-    return (
-      <OutletContext.Provider value={context}>{outlet}</OutletContext.Provider>
-    );
-  }
-  return outlet;
-}
-
-/**
- * Returns an object of key/value pairs of the dynamic params from the current
- * URL that were matched by the route path.
- *
- * @see https://reactrouter.com/docs/en/v6/api#useparams
- */
-export function useParams<
-  ParamsOrKey extends string | Record<string, string | undefined> = string
->(): Readonly<
-  [ParamsOrKey] extends [string] ? Params<ParamsOrKey> : Partial<ParamsOrKey>
-> {
-  let { matches } = React.useContext(RouteContext);
-  let routeMatch = matches[matches.length - 1];
-  return routeMatch ? (routeMatch.params as any) : {};
-}
-
-/**
- * Resolves the pathname of the given `to` value against the current location.
- *
- * @see https://reactrouter.com/docs/en/v6/api#useresolvedpath
- */
-export function useResolvedPath(to: To): Path {
-  let { matches } = React.useContext(RouteContext);
-  let { pathname: locationPathname } = useLocation();
-
-  let routePathnamesJson = JSON.stringify(
-    matches.map((match) => match.pathnameBase)
-  );
-
-  return React.useMemo(
-    () => resolveTo(to, JSON.parse(routePathnamesJson), locationPathname),
-    [to, routePathnamesJson, locationPathname]
-  );
-}
-
-/**
- * Returns the element of the route that matched the current location, prepared
- * with the correct context to render the remainder of the route tree. Route
- * elements in the tree must render an <Outlet> to render their child route's
- * element.
- *
- * @see https://reactrouter.com/docs/en/v6/api#useroutes
- */
-export function useRoutes(
-  routes: RouteObject[],
-  locationArg?: Partial<Location> | string
-): React.ReactElement | null {
-  invariant(
-    useInRouterContext(),
-    // TODO: This error is probably because they somehow have 2 versions of the
-    // router loaded. We can help them understand how to avoid that.
-    `useRoutes() may be used only in the context of a <Router> component.`
-  );
-
-  let { matches: parentMatches } = React.useContext(RouteContext);
-  let routeMatch = parentMatches[parentMatches.length - 1];
-  let parentParams = routeMatch ? routeMatch.params : {};
-  let parentPathname = routeMatch ? routeMatch.pathname : "/";
-  let parentPathnameBase = routeMatch ? routeMatch.pathnameBase : "/";
-  let parentRoute = routeMatch && routeMatch.route;
-
-  if (__DEV__) {
-    // You won't get a warning about 2 different <Routes> under a <Route>
-    // without a trailing *, but this is a best-effort warning anyway since we
-    // cannot even give the warning unless they land at the parent route.
-    //
-    // Example:
-    //
-    // <Routes>
-    //   {/* This route path MUST end with /* because otherwise
-    //       it will never match /blog/post/123 */}
-    //   <Route path="blog" element={<Blog />} />
-    //   <Route path="blog/feed" element={<BlogFeed />} />
-    // </Routes>
-    //
-    // function Blog() {
-    //   return (
-    //     <Routes>
-    //       <Route path="post/:id" element={<Post />} />
-    //     </Routes>
-    //   );
-    // }
-    let parentPath = (parentRoute && parentRoute.path) || "";
-    warningOnce(
-      parentPathname,
-      !parentRoute || parentPath.endsWith("*"),
-      `You rendered descendant <Routes> (or called \`useRoutes()\`) at ` +
-        `"${parentPathname}" (under <Route path="${parentPath}">) but the ` +
-        `parent route path has no trailing "*". This means if you navigate ` +
-        `deeper, the parent won't match anymore and therefore the child ` +
-        `routes will never render.\n\n` +
-        `Please change the parent <Route path="${parentPath}"> to <Route ` +
-        `path="${parentPath === "/" ? "*" : `${parentPath}/*`}">.`
-    );
-  }
-
-  let locationFromContext = useLocation();
-
-  let location;
-  if (locationArg) {
-    let parsedLocationArg =
-      typeof locationArg === "string" ? parsePath(locationArg) : locationArg;
-
-    invariant(
-      parentPathnameBase === "/" ||
-        parsedLocationArg.pathname?.startsWith(parentPathnameBase),
-      `When overriding the location using \`<Routes location>\` or \`useRoutes(routes, location)\`, ` +
-        `the location pathname must begin with the portion of the URL pathname that was ` +
-        `matched by all parent routes. The current pathname base is "${parentPathnameBase}" ` +
-        `but pathname "${parsedLocationArg.pathname}" was given in the \`location\` prop.`
-    );
-
-    location = parsedLocationArg;
-  } else {
-    location = locationFromContext;
-  }
-
-  let pathname = location.pathname || "/";
-  let remainingPathname =
-    parentPathnameBase === "/"
-      ? pathname
-      : pathname.slice(parentPathnameBase.length) || "/";
-  let matches = matchRoutes(routes, { pathname: remainingPathname });
-
-  if (__DEV__) {
-    warning(
-      parentRoute || matches != null,
-      `No routes matched location "${location.pathname}${location.search}${location.hash}" `
-    );
-
-    warning(
-      matches == null ||
-        matches[matches.length - 1].route.element !== undefined,
-      `Matched leaf route at location "${location.pathname}${location.search}${location.hash}" does not have an element. ` +
-        `This means it will render an <Outlet /> with a null value by default resulting in an "empty" page.`
-    );
-  }
-
-  return _renderMatches(
-    matches &&
-      matches.map((match) =>
-        Object.assign({}, match, {
-          params: Object.assign({}, parentParams, match.params),
-          pathname: joinPaths([parentPathnameBase, match.pathname]),
-          pathnameBase:
-            match.pathnameBase === "/"
-              ? parentPathnameBase
-              : joinPaths([parentPathnameBase, match.pathnameBase]),
-        })
-      ),
-    parentMatches
-  );
-}
-
-///////////////////////////////////////////////////////////////////////////////
 // UTILS
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -716,27 +382,6 @@ export function renderMatches(
   matches: RouteMatch[] | null
 ): React.ReactElement | null {
   return _renderMatches(matches);
-}
-
-function _renderMatches(
-  matches: RouteMatch[] | null,
-  parentMatches: RouteMatch[] = []
-): React.ReactElement | null {
-  if (matches == null) return null;
-
-  return matches.reduceRight((outlet, match, index) => {
-    return (
-      <RouteContext.Provider
-        children={
-          match.route.element !== undefined ? match.route.element : outlet
-        }
-        value={{
-          outlet,
-          matches: parentMatches.concat(matches.slice(0, index + 1)),
-        }}
-      />
-    );
-  }, null as React.ReactElement | null);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
